Initial commit
Adds a frontend and backend for this app. Uses the frontend to
either accept English contents for translation or direct LLM usage
to generate content. When a user enters content to be translated,
it is sent to DeepL to be translated for later LLM usage.

The LLM either takes that content or a user prompt. The system
prompt is to translate content to Brazillian Portuguese at a B1
level. So either taking the translated content or a prompt, creates
Portuguese content at my level. I can click on any of the sentences
and if confirmed, it will send a request to Anki to add and sync the
card. This allows me to drill flashcards on Desktop or mobile, wherever
I have the app installed.